requirements.txt
fastapi==0.103.2
uvicorn[standard]==0.23.2
motor==3.2.0
python-decouple==3.8
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic[email]==1.10.12
bcrypt==4.0.1
